back-end/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.0.4
bcrypt==4.1.2
billiard==4.2.0
build==1.1.1
CacheControl==0.14.0
celery==5.3.6
certifi==2023.11.17
channels==4.0.0
charset-normalizer==3.3.2
check-manifest==0.49
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.4.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==23.5
django-import-export==3.3.7
django-redis==5.4.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.5.0
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.3.4
dulwich==0.21.7
et-xmlfile==1.1.0
fastjsonschema==2.19.1
filelock==3.13.1
ghp-import==2.1.0
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
install==1.3.5
installer==0.7.0
itypes==1.2.0
jaraco.classes==3.3.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kavenegar==1.1.2
keyring==24.3.1
kombu==5.3.5
Markdown==3.5.2
MarkupPy==1.14
MarkupSafe==2.1.5
mergedeep==1.3.4
mkdocs==1.5.3
more-itertools==10.2.0
msgpack==1.0.8
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==24.0
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.10.0
platformdirs==4.2.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
promise==2.3
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pycodestyle==2.11.1
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-memcached==1.59
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyyaml_env_tag==0.1
rapidfuzz==3.6.2
redis==5.0.2
referencing==0.34.0
requests==2.31.0
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rpds-py==0.18.0
setuptools==69.0.3
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sql-judge-utils==0.1.8
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
tomlkit==0.12.4
trove-classifiers==2024.3.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
virtualenv==20.25.0
watchdog==4.0.0
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
